Fix "Diamond of death" bug in translation machinery
It was possible to trigger compilation of formulae that were listed as
`:build` dependencies if there was a complex relationship between the
top-level dependencies and the lower ones.  We fix this by adding to the
translation machinery a simple regex replacement that replaces depends_on
statements with previously translated formulae paths, simplifying the
complex relationships.